# api/openapi.yml
openapi: 3.0.1
info:
  title: House of Bills API
  version: v0
servers:
  - url: http://localhost:8080/api/v1
    description: House of Bills API
paths:
...
  /houses:
    post:
      tags:
        - Houses
      operationId: createHouse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseRequest'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/House'
  /houses/batch:
    post:
      tags:
        - Houses
      operationId: createHousesBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseBatchRequest'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/House'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /houses/{id}:
    get:
      tags:
        - Houses
      operationId: getHouseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/House'
        404:
          description: Not Found
    delete:
      tags:
        - Houses
      operationId: deleteHouseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
        404:
          description: Not Found
    put:
      tags:
        - Houses
      operationId: updateHouseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHouseRequest'
      responses:
        200:
          description: Ok
        404:
          description: Not Found
  /houses/user/{id}:
    get:
      tags:
        - Houses
      operationId: getHousesByUserId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/House'
        404:
          description: Not Found
  /incomes:
    post:
      tags:
        - Incomes
      operationId: createIncome
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncomeRequest'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Income'
  /incomes/batch:
    post:
      tags:
        - Incomes
      operationId: createIncomesBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncomeBatchRequest'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Income'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /incomes/{id}:
    get:
      tags:
        - Incomes
      operationId: getIncomeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Income'
        404:
          description: Not Found
  /incomes/house/{id}:
    get:
      tags:
        - Incomes
      operationId: getIncomesByHouseId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 25
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Income'
        404:
          description: Not Found
...
components:
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerId:
          type: string
          format: uuid
...
    CreateHouseRequest:
      type: object
      properties:
        name:
          type: string
        countryCode:
          type: string
        city:
          type: string
        streetLine1:
          type: string
        streetLine2:
          type: string
        userId:
          type: string
          format: uuid
        groupIds:
          type: array
          items:
            type: string
            format: uuid
    House:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        countryCode:
          type: string
        city:
          type: string
        streetLine1:
          type: string
        streetLine2:
          type: string
        userId:
          type: string
          format: uuid
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    CreateHouseBatchRequest:
      type: object
      properties:
        houses:
          type: array
          items:
            $ref: '#/components/schemas/CreateHouseRequest'
    CreateIncomeRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        sum:
          type: number
        houseId:
          type: string
          format: uuid
        groupIds:
          type: array
          items:
            type: string
            format: uuid
    Income:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        sum:
          type: number
        houseId:
          type: string
          format: uuid
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    CreateIncomeBatchRequest:
      type: object
      properties:
        incomes:
          type: array
          items:
            $ref: '#/components/schemas/CreateIncomeRequest'
...
    Error:
      type: object
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: string